backend/lambdas/extraction_worker/requirements.txt
boto3>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
google-generativeai>=0.3.0
openai>=1.12.0
pandas>=2.0.0
python-docx>=0.8.11
Pillow>=10.0.0
PyPDF2>=3.0.0
